requirements.txt
flask
pymongo
<<<<<<< HEAD
python-dotenv
=======
python-dotenv
>>>>>>> 3460e82d14719d8dd6f27c8016e4aa7d40b32e1a
